LayoutTests/imported/w3c/web-platform-tests/css/css-animations/parsing/animation-shorthand-expected.txt

CONSOLE MESSAGE: line 14: ReferenceError: Can't find variable: test_shorthand_value

FAIL CSS Animations: animation sets longhands ReferenceError: Can't find variable: test_shorthand_value
